requirements.txt
fastapi
uvicorn
mysql-connector-python
joblib
numpy
scikit-learn
gunicorn
